Extract PushDownnegationsExpressionRewriter
